feat: add pickFile() and modernize PlatformFile data access
- Add `pickFile()` static method as a convenience wrapper for single-file selection.
- Add `readAsBytes()` and `readAsByteStream()` to `PlatformFile` to allow loading file data on demand.
- Deprecate `withData`, `withReadStream`, and `readSequential` parameters in `pickFiles()`.
- **BREAKING CHANGE**: Make `fileName` a required parameter in `saveFile()` across all platforms.
- **BREAKING CHANGE**: Update `pickFiles()` to default `allowMultiple` to `true` and deprecate the parameter.
- Refactor iOS file name building and extension inference logic into `FilePickerUtils`.
- Update documentation, README, and example app to reflect new API patterns.
