+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 ### General
 - Offloaded file I/O operations to background threads and isolates to improve UI responsiveness.
 - Removed redundant comments and internal implementation notes across multiple files.
-- Updated `kotlinx-coroutines-android` dependency to version `1.11.0`.
+- Added `pickFile()` static method as a convenience wrapper for single file selection, returning `PlatformFile?` directly. [#1469](https://github.com/miguelpruivo/flutter_file_picker/issues/1469)
+- **BREAKING CHANGE**: Refactored `saveFile()` to make `fileName` and `bytes` required parameters across all platforms for a more consistent API. Improved the method's documentation for better clarity.
+- Improved documentation for `PlatformFile` properties (`path`, `bytes`, `readStream`) to clarify nullability and usage across platforms. [#1469](https://github.com/miguelpruivo/flutter_file_picker/issues/1469)
+- **BREAKING CHANGE**: The `allowMultiple` parameter on `pickFiles()` now defaults to `true`. Use `pickFile()` for single-file selection. (`allowMultiple` is deprecated and will be removed in a future release.)
+- Deprecated `withData`, `withReadStream`, and `readSequential` on `pickFiles()`/`pickFile()`. Users should call `PlatformFile.readAsBytes()` or `PlatformFile.readAsByteStream()` to load file data on demand. These parameters will be removed in a future release.
 
 ### Android
-- Added `kotlinx-coroutines-android` dependency to handle asynchronous operations efficiently.
-- Removed redundant null bytes entry from `FileInfo` map.
 - Fixed an issue where `FileType.any` would prevent subdirectories from being listed in the system file explorer by ensuring `EXTRA_MIME_TYPES` is correctly passed as an array. [#2013](https://github.com/miguelpruivo/flutter_file_picker/issues/2013)
+- Updated `kotlinx-coroutines-android` dependency to version `1.11.0`.
 
 ## 12.0.0-beta.3
 ### General
