+++ analysis_options.yaml
@@ -1,1 +1,5 @@
 include: package:lints/recommended.yaml
+
+analyzer:
+  errors:
+    deprecated_member_use_from_same_package: ignore
