+++ lib/src/api/platform_file.dart
@@ -40,7 +40,10 @@ class PlatformFile {
   /// ```
   /// final File myFile = File(platformFile.path);
   /// ```
-  /// On web the path points to a Blob URL, if present, which can be cleaned up using [URL.revokeObjectURL](https://pub.dev/documentation/web/latest/web/URL/revokeObjectURL.html).
+  ///
+  /// This property is `null` on Android, when using SAF without caching enabled.
+  ///
+  /// On the web this may or may not point to a Blob URL, which can be cleaned up using [URL.revokeObjectURL](https://pub.dev/documentation/web/latest/web/URL/revokeObjectURL.html).
   /// Read more about it [here](https://github.com/miguelpruivo/flutter_file_picker/wiki/FAQ)
   final String? path;
 
@@ -49,10 +52,23 @@ class PlatformFile {
 
   /// Byte data for this file. Particularly useful if you want to manipulate its data
   /// or easily upload to somewhere else.
+  ///
+  /// This property is `null` unless [FilePicker.pickFiles] (or `pickFile`) was called with
+  /// `withData: true`. Enabling this on mobile platforms for large files may lead to
+  /// Out Of Memory (OOM) issues; use [readStream] instead.
+  ///
   /// [Check here in the FAQ](https://github.com/miguelpruivo/flutter_file_picker/wiki/FAQ) an example on how to use it to upload on web.
+  @Deprecated(
+    'Use readAsBytes() instead to avoid out-of-memory issues with large files.',
+  )
   final Uint8List? bytes;
 
-  /// File content as stream
+  /// File content as a stream of bytes.
+  ///
+  /// This property is `null` unless [FilePicker.pickFiles] (or `pickFile`) was called with
+  /// `withReadStream: true`. This is the recommended way to handle large files on
+  /// mobile and desktop platforms.
+  @Deprecated('Use readAsByteStream() instead')
   final Stream<List<int>>? readStream;
 
   /// The file size in bytes. Defaults to `0` if the file size could not be
@@ -62,6 +78,10 @@ class PlatformFile {
   /// The platform identifier for the original file, refers to an [Uri](https://developer.android.com/reference/android/net/Uri) on Android and
   /// to a [NSURL](https://developer.apple.com/documentation/foundation/nsurl) on iOS.
   /// Is set to `null` on all other platforms since those are all already referencing the original file content.
+  ///
+  /// This property is `null` in the following cases:
+  /// - On Web, where local file paths are not available (it may point to a Blob URL instead).
+  /// - On Android, when picking a directory or using SAF without caching enabled.
   ///
   /// Note: You can't use this to create a Dart `File` instance since this is a safe-reference for the original platform files, for
   /// that the [path] property should be used instead.
@@ -73,9 +93,51 @@ class PlatformFile {
   /// Retrieves this as a XFile
   XFile get xFile {
     if (kIsWeb) {
+      if (bytes == null) {
+        throw StateError(
+          'Cannot create XFile on web without bytes. '
+          'Call pickFiles() or pickFile() with withData: true.',
+        );
+      }
       return XFile.fromData(bytes!, name: name, length: size);
     } else {
+      if (path == null) {
+        throw StateError(
+          'Cannot create XFile without a path. '
+          'On Android SAF files without cache, use withData: true when picking.',
+        );
+      }
       return XFile(path!, name: name, bytes: bytes, length: size);
+    }
+  }
+
+  /// Read the file content as bytes.
+  ///
+  /// For large files on mobile and desktop platforms, prefer [readAsByteStream]
+  /// to avoid Out Of Memory (OOM) issues.
+  Future<Uint8List> readAsBytes() => xFile.readAsBytes();
+
+  /// Read the file content as a stream of bytes.
+  ///
+  /// This is the recommended way to handle large files on mobile and desktop platforms.
+  ///
+  /// **Note on Web:** This method currently requires [FilePicker.pickFiles] or [FilePicker.pickFile]
+  /// to have been called with `withData: true`, otherwise this will fail since `bytes` is required
+  /// to create the stream. This is a limitation that should be addressed in a future version to
+  /// support proper streaming on web without requiring `withData`.
+  Stream<Uint8List> readAsByteStream() => xFile.openRead();
+
+  /// Returns the length of the file in bytes.
+  ///
+  /// Note: on Web, `xFile.length()` depends on the underlying `XFile` implementation
+  /// and may require `withData: true` when using `XFile.fromData` — in that case the
+  /// fallback to `bytes` will be used.
+  Future<int> length() async {
+    if (size > 0) return size;
+    try {
+      return await xFile.length();
+    } catch (_) {
+      return bytes?.lengthInBytes ?? 0;
     }
   }
