+++ lib/src/platform/file_picker_method_channel.dart
@@ -171,7 +171,7 @@ class MethodChannelFilePicker extends FilePickerPlatform {
   @override
   Future<String?> saveFile({
     String? dialogTitle,
-    String? fileName,
+    required String fileName,
     String? initialDirectory,
     FileType type = FileType.any,
     List<String>? allowedExtensions,
