+++ lib/src/platform/file_picker_platform_interface.dart
@@ -84,7 +84,7 @@ abstract class FilePickerPlatform extends PlatformInterface {
   /// name to save a file.
   Future<String?> saveFile({
     String? dialogTitle,
-    String? fileName,
+    required String fileName,
     String? initialDirectory,
     FileType type = FileType.any,
     List<String>? allowedExtensions,
