+++ lib/src/platform/macos/file_picker_macos.dart
@@ -99,7 +99,7 @@ class FilePickerMacOS extends FilePickerPlatform {
   @override
   Future<String?> saveFile({
     String? dialogTitle,
-    String? fileName,
+    required String fileName,
     String? initialDirectory,
     FileType type = FileType.any,
     List<String>? allowedExtensions,
