+++ lib/src/platform/windows/file_picker_windows.dart
@@ -180,7 +180,7 @@ class FilePickerWindows extends FilePickerPlatform {
   @override
   Future<String?> saveFile({
     String? dialogTitle,
-    String? fileName,
+    required String fileName,
     String? initialDirectory,
     FileType type = FileType.any,
     List<String>? allowedExtensions,
